+++ brain-react/section_07/src/components/templates/TodoCreateTemplate/index.jsx
@@ -3,14 +3,12 @@
  *
  * @package components
  */
-import { useState, useCallback } from "react";
-import { useNavigate } from "react-router-dom";
 import { BaseLayout } from "../../organisims/BaseLayout";
 import { InputForm } from "../../atoms/InputForm";
 import { TextArea } from "../../atoms/TextArea";
 import { CommonButton } from "../../atoms/CommonButton";
 import { useTodoContext } from "../../../contexts/TodoContext";
-import { NAVIGATION_LIST } from "../../../constants/navigations";
+import { useTodoCreateTemplate } from "./useTopCreateTemplate";
 import styles from "./styles.module.css";
 
 /**
@@ -20,45 +18,11 @@ import styles from "./styles.module.css";
  * @constructor
  */
 export const TodoCreateTemplate = () => {
-    const navigate = useNavigate();
     const { addTodo } = useTodoContext();
-    /* local state */
-    const [inputTitle, setInputTitle] = useState("");
-    const [inputContent, setInputContent] = useState("");
-
-    /**
-     * 「title」の変更処理
-     *
-     * @type {function(*): void}
-     */
-    const handleChangeTitle = useCallback(
-        (e) => setInputTitle(e.target.value), []
-    );
-
-    /**
-     * 「content」変更処理
-     *
-     * @type {function(*): void}
-     */
-    const handleChangeContent = useCallback(
-        (e) => setInputContent(e.target.value), []
-    );
-
-    /**
-     * Todo追加処理
-     *
-     * @type {function(*): void|*}
-     */
-    const handleCreateTodo = useCallback(
-        (e) => {
-            e.preventDefault();
-            if (inputTitle !== "" && inputContent !== "") {
-                addTodo(inputTitle, inputContent);
-                navigate(NAVIGATION_LIST.TOP);
-            }
-        },
-        [addTodo, inputTitle, inputContent, navigate]
-    );
+    const [
+        { inputTitle, inputContent },
+        { handleChangeTitle, handleChangeContent, handleCreateTodo }
+    ] = useTodoCreateTemplate({ addTodo });
 
     return (
         <BaseLayout title={"Create Todo"}>
