+++ brain-react/section_07/src/components/templates/TodoCreateTemplate/useTopCreateTemplate.js
@@ -1,0 +1,68 @@
+/**
+ * useTodoCreateTemplate
+ *
+ * @package components
+ */
+import { useState, useCallback } from "react";
+import { useNavigate } from "react-router-dom";
+import { NAVIGATION_LIST } from "../../../constants/navigations";
+
+/**
+ * useTodoCreateTemplate
+ *
+ * @param addTodo
+ */
+export const useTodoCreateTemplate = ({ addTodo }) => {
+    const navigate = useNavigate();
+
+    /* local state */
+    const [inputTitle, setInputTitle] = useState("");
+    const [inputContent, setInputContent] = useState("");
+
+    /**
+     * 「title」の変更処理
+     *
+     * @type {function(*): void}
+     */
+    const handleChangeTitle = useCallback(
+        (e) => setInputTitle(e.target.value), []
+    );
+
+    /**
+     * 「content」変更処理
+     *
+     * @type {function(*): void}
+     */
+    const handleChangeContent = useCallback(
+        (e) => setInputContent(e.target.value), []
+    );
+
+    /**
+     * Todo追加処理
+     *
+     * @type {function(*): void|*}
+     */
+    const handleCreateTodo = useCallback(
+        (e) => {
+            e.preventDefault();
+            if (inputTitle !== "" && inputContent !== "") {
+                addTodo(inputTitle, inputContent);
+                navigate(NAVIGATION_LIST.TOP);
+            }
+        },
+        [addTodo, inputTitle, inputContent, navigate]
+    );
+
+    const states = {
+        inputTitle,
+        inputContent
+    };
+
+    const actions = {
+        handleChangeTitle,
+        handleChangeContent,
+        handleCreateTodo
+    };
+
+    return [states, actions];
+};
